got the overall qt4 gui to open and run on ubuntu 22.04, for pull request #530
The "QPixmap: Must construct a QApplication before a QPaintDevice" seg fault error went away by commenting out the "aboutQtAction" stuff in mainWindow.cpp, but then the gui would hang without running until I commented out anything and everything related to the "QFutureWatcher" class. So the "futureWatcher" stuff in WidgetDownloadDEM and stationFetchWidget, which seem to be related to the progress bars and data downloads of the download dem and pointInit pages.

In addition, opening the dem download page gives a bunch of OpenSSL within qt4 errors, "QSslSocket: cannot call unresolved function SSL_get_peer_certificate". I tried to fix them by calling "QSslSocket::addDefaultCaCertificates("path_to_ssl.crt");", but then I got a huge bunch of errors which I was not able to resolve, that seemed to imply that this qt4 was not built with any openssl functionality. Running ldd on WindNinja shows that it is grabbing the system openssl which is version 3, so it could just be version conflicts of openssl, but when I hand built openssl 1.0.1u, cmake would find the properly updated openssl stuff but ldd would not even find the proper openssl stuff even when setting LD_LIBRARY_PATH, so I had even worse version conflicts. In the end, it seemed like the real problem was that qt4 itself didn't have the proper openssl stuff built in, implying that even pointing WindNinja to the proper openssl version would not be enough, qt4 itself would have to be rebuilt with openssl somehow which is not an option.

With these fixes and problems in mind, the gui opens fine now, can click on and do whatever in WindNinja as far as I can tell, including a domain avg run with all outputs and a weather model download run. However, the download dem page doesn't display the map, and both the download dem and the pointInit data downloads just sit there hanging instead of downloading their files. So these are major problems that would need to be fixed.

In addition, I still need to test and make sure these changes don't break windninja builds on other systems. Also, to find these fixes I had commented out anything gui related then slowly added gui stuff back in, adding additional headers where needed. But as can be seen with this commit, the additional headers weren't needed when all the code was put back together at once just commenting out the problem code. So in future commits, I may add some of the additional headers to the code that I had to add when breaking things into parts, to make the code a bit more precise. Also, it kind of seemed to me like the GoogleMapsInterface and latLonWidget were unused bits of gui code, might be good to see if they can be dropped from the gui code in the future.